api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-docx==1.1.2
google-api-python-client==2.141.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
